[Exp PyROOT][ROOT-8011] Turn libJupyROOT into a C extension module
As discussed in ROOT-8011, TPython can't work with the ROOT C++
kernel for Jupyter because the current implementation relies
on a plain shared library to run the cell code and capture
its output and error. In particular, a crash happens when
the library, libJupyROOT, tries to get the __main__ module
in TPython's initialization.

Turning libJupyROOT into a C extension module makes TPython's
initialization succeed, since Python is fully operational.