[CMake] python-dev required only for PyROOT
python-dev is required to build PyROOT with multiple Python versions.
The reason is that the python related variables are set only if
python-dev was found. On the other hand, python-dev is not a requirement
for ROOT per se, since it just requires interpreter.

With this change we relax the condition on python-dev found and force
pyroot and pyroot_exp to off it python-dev was not found.