[DF] Keep objects used by jitted code alive until jitting time
RJittedFilter and RJittedCustomColumn objects pertaining to jitted
filters or defines that were registered and then never used could
potentially go out of scope before jitting happened. This is now
prevented by storing a shared_ptr to these objects on the heap, that
jitted code deletes when run.

This fixes ROOT-10619.