[Exp PyROOT] Implement Address via LowLevelView
The previous implementation, based on a bytearray, did not consider
that, in the old PyROOT, the address of the buffer returned by
AddressOf is the same as the address of the address of the object
being inspected. This is required by DataFormats.FWLite in CMSSW.

In the new implementation, we obtain the address of the address of
the object via addressof (and its parameter byref = True), and
then we just create a cppyy LowLevelView from it, which is the
buffer we return. The first (and only) element of that buffer
contains the address of the object.