[treeplayer] Try harder to find leaf count (ROOT-10397):
Before, the leaf count was assumed to exist at the same level as the leaf to be read,
e.g. a.b.x[n] would look for a.b.n. That fails if n is coming from a "higher level".
Now, TUIntOrIntReader also looks for a.n and n, if a.b.n is not found.

And add the repro test for this.