Move changes to TTreeCache data members in apparently better positions (#5365)
* Move changes to TTreeCache data members in apparently better positions

These minor changes seem to affect the way in which TTreeCache interacts with different IO classes, e.g. TTree/TChain vs TTreeReader.

In `TTreeCache::FillBuffer` the change in values of `fEntryCurrent` and `fEntryNext`, respectively the first entry of the current cluster and the first entry of the next cluster, has been moved before an `if` condition that was causing an early exit from the function.
This, in conjunction with the change in `TTreeReader::SetEntriesRange` apparently fixes the issues with the cache when running a simple program that uses only TTreeReader.

At the same time, the changes affect the interaction of the cache with a TTree-only program in a way that now is throwing a log error:
```Error in <TTreeCache::FillBuffer>: Inconsistency: fCurrentClusterStart=821695 fEntryCurrent=821695 fNextClusterStart=1643390 but fEntryCurrent should not be in between the two```

I'm not sure why these changes affect positively the TTreeReader examples and negatively the TTree ones. Some links to these examples follow.

* Start learning phase from minimum entry in the cache
When the cache is in the learning phase, the entry to start filling the buffer from should be set to `fEntryMin`, whereas before it was simply set to zero. This meant that for entry ranges outside of the first cluster the cache was still trying to fill buffers from the beginning of the tree. With the previous commit to `TTreeReader` this was also causing early returns of the `FillBuffer` function, resulting in an infinite loop over the first cluster with no cache.

This behaviour affects specifically the file-based caching mechanism offered by `TFilePrefetch`, since resetting the entry to zero happens inside an `if (fEnablePrefetching)` statement.