[DF] Fix memleaks of dataframe nodes kept alive for jitted code
Before this patch, if a RLoopManager went out of scope before running
its event loop (a rare occurrence, but indeed possible, especially in
interactive sessions), we were leaking certain weak_ptrs and shared_ptrs
to components of the computation graph, that were meant to be used by
lazily jitted code that was never actually jitted in the end.

With this change, when _any_ RLoopManager runs the event loop, we
trigger execution of _all_ jitted code that was registered, even if
it was registered by an RLoopManager that is now out of scope. In the
latter case, the code only performs clean-up of the heap-allocated
weak_ptrs and shared_ptrs and exits. We still "leak" in the odd case
in which RDF jitted code is registered but the application never actually
triggers any RDF event loops, in which case the code to be lazily jitted
remains in the pipeline until the end of the application, and parts of
the computation graph are kept alive indefinitely by heap-allocated shared_ptrs.