[CMake][Exp PyROOT] Split CMake Python variables in ROOT and PyROOT/PyMVA related
These changes address the case in which Python 3 has only Interpreter
while Python 2 has both Interpreter and Development. In this case, the
Python used to build ROOT (which requires only Interpreter) will be 3,
but PyROOT (which requires also Development) will be built only for 2.
To achieve this, ROOT and PyROOT/PyMVA use now two different sets of
variables.

Documentation for the whole machinery is also added.

It is also worth pointing out that the entire machinery could be much
simplified by just requiring the following as a prerequisite for ROOT:
        - CMake >= 3.12.4
        - Python-Development package