[PyROOT] Conditionally add thread to process ROOT events
The PR spawns a separate thread to process ROOT events if we cannot
attach the update mechanism to an exisiting hook such as existing in the
interactive Python mode or for ipython/notebooks.

Also adds backward compatibility for ROOT.PyROOT.StartGUIThread and
shuts down the thread safely at teardown time.

Related to ROOT-10774