Add CPack options for DEB/RPM package generation
This adds the necessary CPack variables to allow package generation
in DEB and RPM formats. It should be noted that in order to generate
automatic library and package dependencies for both DEB and RPM, no
builtins should be enabled. Using /usr as prefix and setting proper
values for CMAKE_INSTALL_LIBDIR and CMAKE_INSTALL_PYTHONDIR in
particular is also strongly encouraged, so that no further setup is
required for using ROOT after installation with apt or dnf. Typically,
that means setting LIBDIR to /usr/lib/x86_64-linux-gnu on Debian and
Ubuntu, and PYTHONDIR to /usr/lib/python3/dist-packages. For RPM-based
distros, like CentOS and RedHat LIBDIR should be set to /usr/lib64 and
PYTHONDIR should be set to e.g. /usr/lib64/python3.6/site-packages.

The full recommended settings for Debian/Ubuntu (for x86_64) are

-Dgnuinstall=ON
-Dsoversion=ON
-DCMAKE_INSTALL_PREFIX=/usr
-DCMAKE_INSTALL_BINDIR=bin
-DCMAKE_INSTALL_CMAKEDIR=lib/x86_64-linux-gnu/cmake/ROOT
-DCMAKE_INSTALL_DATAROOTDIR=share
-DCMAKE_INSTALL_DATADIR=share/root
-DCMAKE_INSTALL_DOCDIR=share/doc/root
-DCMAKE_INSTALL_ELISPDIR=share/emacs/site-lisp
-DCMAKE_INSTALL_FONTDIR=share/root/fonts
-DCMAKE_INSTALL_ICONDIR=share/root/icons
-DCMAKE_INSTALL_INCLUDEDIR=include/ROOT
-DCMAKE_INSTALL_JSROOTDIR=share/root/js
-DCMAKE_INSTALL_LIBDIR=lib/x86_64-linux-gnu
-DCMAKE_INSTALL_MACRODIR=share/root/macros
-DCMAKE_INSTALL_MANDIR=share/man
-DCMAKE_INSTALL_OPENUI5DIR=share/root/ui5
-DCMAKE_INSTALL_PYTHONDIR=lib/python3/dist-packages
-DCMAKE_INSTALL_SRCDIR=/dev/null
-DCMAKE_INSTALL_SYSCONFDIR=/etc/root
-DCMAKE_INSTALL_TUTDIR=share/root/tutorials

and for CentOS/RedHat the recommended settings are

-Dgnuinstall=ON
-Dsoversion=ON
-DCMAKE_INSTALL_PREFIX=/usr
-DCMAKE_INSTALL_BINDIR=bin
-DCMAKE_INSTALL_CMAKEDIR=lib64/cmake/ROOT
-DCMAKE_INSTALL_DATAROOTDIR=share
-DCMAKE_INSTALL_DATADIR=share/root
-DCMAKE_INSTALL_DOCDIR=share/doc/root
-DCMAKE_INSTALL_ELISPDIR=share/emacs/site-lisp
-DCMAKE_INSTALL_FONTDIR=share/root/fonts
-DCMAKE_INSTALL_ICONDIR=share/root/icons
-DCMAKE_INSTALL_INCLUDEDIR=include/ROOT
-DCMAKE_INSTALL_JSROOTDIR=share/root/js
-DCMAKE_INSTALL_LIBDIR=lib64
-DCMAKE_INSTALL_MACRODIR=share/root/macros
-DCMAKE_INSTALL_MANDIR=share/man
-DCMAKE_INSTALL_OPENUI5DIR=share/root/ui5
-DCMAKE_INSTALL_PYTHONDIR=lib64/python3.6/site-packages
-DCMAKE_INSTALL_SRCDIR=/dev/null
-DCMAKE_INSTALL_SYSCONFDIR=/etc/root
-DCMAKE_INSTALL_TUTDIR=share/root/tutorials

If ROOT is configured as shown above, and all dependencies are
installed at build time, the required runtime dependencies will be
generated properly by CPack and will be installed alongside ROOT
when installing the package with apt or rpm. Extra dependencies that
may be required at runtime, like numpy, numba, etc, should be added
individually to the respective CPack variables for recommended and
suggested packages. A few have been added in this commit already.