[v7, ntuple] allow RResult to work with move-only types
Without being able to move from the inner field, we cannot use RResult
to hold move-only types.

RResult::Get() conceptually transfers ownership of the inner value out
from the RResult, so it makes sense to extract the value after a Get
call. The Result is left in an invalid state, but this also makes sense
in the context of ownership. By calling Get(), we "open the box" to see
if the operation succeeded. If it didn't, we throw an exception but if
it succeeded, then there's no need for the Result to stick around because the
inner value was what we wanted anyways.