Add CMAKE_INSTALL_PYTHONDIR build option
Having an extra option makes it possible to customize the installation
directory of PyROOT & friends rather than forcing users/packagers
to install PyROOT along the C++ libraries, which in turn makes it
necessary to have a non-standard PYTHONPATH set otherwise ROOT.py is
not found.

With CMAKE_INSTALL_PYTHONDIR, ROOT's python components can be installed
in standard python-related system directories.

At least as a first step, we limit the new feature to the case of new PyROOT,
non-Windows platforms and single-python builds.