[DF] Add all spellings of a branch name to valid column names
Before this patch, in cases in which t.GetBranch("a.b") and
t.GetBranch("b") both returned a valid address, RDataFrame was adding
only "a.b" to the list of valid TTree columns.
With this patch, both "a.b" and "b" are recognized as valid column names.

We need this change in behavior to avoid a _worse_ change in behavior,
described in detail in ROOT-10942: since ROOT-10702 was fixed,
TTree::GetBranch became more powerful and started returning non-null
addresses for branch names with form "a.b" while it was returning a
nullptr until v6.20/06. With RDataFrame's previous logic, this in turn
meant that valid code that was using "a" as a column broke as RDataFrame
was now adding the "a.b" spelling to the list of valid columns instead.

This fixes the RDF-related part of ROOT-10942: "a" is recognized as a
valid spelling again, and "a.b" is kept as a new valid spelling.