- Fix using correct minimizer when computing the profiled nuisance parameters in the FrequentistCalculator
The Frequentist calculator needs to compute , before generating the toys, the profiled values of the nuisances for the nukll and alt case. RooProfileLL is used but the default (Minuit) minimizer was used.
Change now to use the static default value, which can be changed and also give possibility to control printlevel.
With this change we can now consistently use stressRooStats with Minuit2 minimizer

- Increase also tolerance of stressRooStats tests when using toys. This is needed in case a different minimizer is used