[DF] Fix recently introduced crashes with gcc 7.5.0
A recent refactoring in RDF introduced an incompatibility with gcc
7.5.0: clang and that gcc version disagree on the address of the
std::map argument passed to InitColumnReaders.
RCustomColumn<...>::InitSlot, the _jitted_ caller, sees the correct
address, but InitColumnReaders<...>, the _compiled_ callee, sees a
different address.

For reasons known only to gcc 7.5.0, reducing the number of function
arguments (by aggregating some of them in a helper struct) resolves
the ABI incompatibility.
Valgrind could see an access to invalid memory before, and it does
not anymore.