[metacling] Remove test dependency on llvm:
One set of llvm used to get linked into the test; a second set got loaded through
libCling. That is fine as long as libCling does not export symbols. If it does, on
the other hand (e.g. for cling-test), static initialization fails.

This test used llvm::Path and StringView out of convenience; it was not an integral
part of the test, and replacing it is well worth the linkage simplification.

We will need to think of a test that
- checks the experiment scenario of loading a custom llvm library besides ROOT, and
- survives cling-test.