[PyROOT][6467] For Python-C++ derived types, invoke constructor via the dispatcher
The previous approach, based on comparing a string containing the
argument types to the signature of the constructor, did not work
in presence of default arguments (see 6467). The new approach
reuses the machinery in place to try the constructors of Dispatcher.

On the other hand, a constructor that accepted a PyObject* is no
longer added to the Dispatcher class, since it was triggering
a corruption of the dispatcher pointer when being attempted
in the reproducer in 6467.

These changes synchronize this part of the code with what is
present in upstream cppyy.