[PyROOT][6468] Adapt SetBranchAddress pythonization to caching of data members
If a class branch has an array data member, cppyy applies an
optimization to cache it and therefore optimize its access in
a loop. Such feature lazily changes the address of the class object
on first access to the data member, so the address previously
used in the pythonization of SetBranchAddress is not correct anymore,
which causes a crash in subsequent invocations of GetEntry.

This fix early triggers the caching in the pythonization of
SetBranchAddress, so that the address will stay correct.
Moreover, the address of the branch object is obtained via
CPPInstance methods that take into account if caching is in place
(ExtendedData for the object has been created).