Use a cache variable for CMAKE_INSTALL_RPATH
This allows the value of this variable to be overriden when calling
CMake to configure ROOT. In particular it will allow release binaries to
be configured with a relative RUNPATH such that setting LD_LIBRARY_PATH
will not be necessary for running ROOT from the binaries, regardless of
where the binary tarballs have been unpacked into the local filesystem.
For non-gnuinstall configuration on Linux, the relative RUNPATH should be
set to '$ORIGIN/../lib' so that it works for both binaries and libraries.
If later on ROOT moves to use lib or lib64 as default depending on the
system, then '$ORIGIN/../$LIB' can be used to make the relative RUNPATH
work in both cases. If LD_LIBRARY_PATH is set anyway, it usually
overrides the embedded RUNPATH set by this method with CMake. See ld(1)
and ld.so(8) for more information on RPATH, RUNPATH, and their precedence
relative to LD_LIBRARY_PATH.