[DF] Write RVecs out from Snapshot as RVecs, not as std::vectors
Now that we use the "collection proxy" mechanism to write RVecs,
they can still be read back as std::vectors or with TTreeReaderArrays
without issues, which removes the motivation for performing this
type-to-type translation on the fly.

Users might notice the change but it should not have any problematic
consequence as RDF and ROOT will still be able to read the RVec
branches written before this commit (without any change in user code
thanks to the "collection proxy" mechanism).