[DF] Delay call to TTree::AddClone until first processed entry
In rare cases, when processing a TChain and producing a Snapshot,
no entry in the first input TTree passes all Filters, so we never
have the chance to create the branches of the output TTree.
As a consequence, when switching from the first file to the next,
TTree::CopyAddresses fails to find the corresponding branches in
the output tree (which was added as its clone via TTree::AddClone),
and prints a warning (that can be safely ignored).

We now instead add the output tree of a Snapshot as a clone of the
input tree more lazily, only when we first process an entry (if ever).
This removes the aforementioned warning, reported at #6848.