Uncomment line setting operation mode of partInt components of RooProdPdf using current operation mode.
This seems to be needed in order to propagate to the cached intergal component s of the RooProdPdf is operation mode.
It fixes issue #7157 where RooProdPdf is used from a RooFFTConv. In this case
the RooFFTConv sets AlwaysDirty as operation mode its components, and then the ROoProdPdf needs to propage to its owned parts, otherwise during the FFT computation the RooProdPdf returns always its first evaluation.