[RF] Adds RooCrystalBall class that fully parametrizes crystal ball
RooFit currently includes the `RooCBShape` and `RooDSCBShape` to
represent the family of crystal ball distributions. The `RooCBShape`
only has a left tail, and `RooDSCBShape` has two independently
parametrized tails.

Some users also want a crystal ball where the width of the Gaussian
core is different on the left and the right side (see e.g. Jira ticket
[ROOT-10388](https://sft.its.cern.ch/jira/browse/ROOT-10388)).

To make the RooFit code more maintainable and also cover the use case
of an asymmetric Gaussian core, this commit implements a new
`RooCrystalBall` class that should cover all classes of crystal ball
shapes, such that `RooCBShape` and `RooDSCBShape` are in principle not
necessary anymore.

The new features of the `RooCrystalBall` class that make it different
from `RooDSCBShape` are:

 - possibility to set sigma of left side and right side of Gaussian
   core separately

 - code modernization and using `std::` math functions

 - three constructors: fully parametrized, symmetric Gaussian core
   (to replace `RooDSCBShape`), only left tail (to replace
   `RooCBShape`)

 - useage of inlined helper functions to avoid code duplication

It is also carefully tested in `testRooCrystalBal`:

 - test the fully parameterized case

 - test the constructor for left-tail only case and compare values
   with `RooCBShape`

 - test the constructor for right-tail only and symmetric tails

 - test the constructor with symmetric Gaussian core and compare
   values with `RooDSCBShape`

 - test integration with many possible combinations integration
   ranges to cover all branches of integration code