[DF] Allow casting RResultPtr<Derived> to RResultPtr<Base>
...and add a test. This fixes ROOT-10321.

This is also the proper fix for ROOT-9785 (basically the same
issue/feature request), which was supposed to be fixed by
78fca0b734fa but unfortunately that commit enabled the conversion
based on the wrong condition: `is_convertible<T, TO>` rather
than `is_convertible<T*, TO*>`.
In order to fully delegate the decision on convertibility to
`shared_ptr`, we now use `is_constructible<shared_ptr<TO>, shared_ptr<T>>`
as the condition instead.