[DF] Remove RResultPtr::Release (an unreleased feature)
The current implementation is broken (fObjPtr.release() deletes the
object thet we then return by pointer) and in general it is extremely
awkward to convince a std::shared_ptr to release ownership (and for good
reason, it only makes sense if you can be absolutely sure that there is
only one shared_ptr alive that owns that object).

Instead of using RResultPtr::Release, we recommend moving the result
out of the RResultPtr.

This effectively reverts commit 03fd58eec04ddf96c1596c02d4269185fef42b9f.