[RF] Ignore leaves in evaluateSpan that the RooAbsReal doesn't depend on
We should not use RooAbsArg::leafNodeServerList to find all leaves,
whose values neet to be set in RooAbsReal::evaluateSpan. Sometimes a
RooAbsReal sits on top of a leaf in the computation graph, still it doesn't
depend on it's values. The example here is a RooRealIntegral, which sets
the leaf values itself to integrate over them. That's why we only add
the parameters and observables here.