[ntuple] Properly handle RPage memory in SinkBuf
RPages do not own their memory. They only have a non-owning pointer,
which will likely be invalidated by the next CommitPage call.

In this case, RColumn manages the RPage memory, and uses it to map in
different pages as necessary. To successfully buffer pages, we have to
copy the memory that the RPage points so we can write it out later.

The previous commits did not take this into account, so when the time
came to flush the buffered pages, each column page would dutifully write
out a copy of the RColumn's head page data.