[RF] Speedup histfactory ParamHistFunc
The ParamHistFunc in the histfactory has a persistent std::map<int,int>
data member. It can be replaced with a std::vector<int> where the
previous key is the index in the vector. This is much faster.

Schema evolution is not a problem here. With this commit, the data
member is made non-persistent because it can be computed from other
persistent data members.