[RF] Implement mechanism in RooCmdArg to take ownership of temporaries
It happens often that one passes a temporary RooArgSet to the
RooCmdArg constructor, where the RooCmdArg is itself a temporary that is
passed to a RooFit function. The RooCmdArg contains only pointers to the
RooArgSets, so this pattern relies on the RooArgSet being kept alive
until the function end.

In code, this pattern would look like this statement:
```C++
func(RooCmdArg{RooArgSet{}});
```

In C++, this works because the temporaries survive until the statement
end. But in Python, the inner temporary RooArgSet will be destructed
after the RooCmdArg has been constructed, leaving the RooCmdArg with
dangling pointers.

To solve this problem, all the functions that can create a RooCmdArg
from a RooArgSet are overloaded with versions accepting rvalue
references. These overloads are indeed used for temporaries both in C++
and Python. In the rvalue reference versions, the temporaries are moved
into a vector of std::unique_ptr<TObject>. The constructed RooCmdArg
then takes ownership of this vector wrapped in a shared_ptr.