[ntuple] enable building with libdaos_mock if `testing` is set
In order to enable RNTuple DAOS unit tests, RNTuple is built with libdaos_mock
if `testing` is set and the user did not enable explicit DAOS support, i.e.
`-Ddaos=ON`.  This allows for testing the DAOS backend in an environment that
lacks a real DAOS deployment.

Note that this is disabled for Windows, as both `libdaos` and `libdaos_mock`
depend on libuuid (which is not tipically there on these platforms).