[ntuple, daos] Defer `daos_init()` call until needed
The RAII object in RDaos.cxx was constructed statically, which causes
a DAOS-enabled RNTuple build to call `daos_init()` (or `daos_fini()` at exit)
even if the only the file backend was used by user code.

This commit turns the RAII object into a static variable in
`RDaosPool::RDaosPool()` which is initialized during the first call to the
constructor.  Note that static variable initialization is guaranteed to be
thread-safe by ISO C++.