[ntuple] perf: Use map::find instead of map::count
RClusterDescriptor::ContainsColumn is called as part of the main read
path and so is a candidate for optimization efforts:

RColumn::MapPage
   RPageSourceFile::PopulatePage
      RNTupleDescriptor::FindClusterId
         RClusterDescriptor::ContainsColumn

A comparison of the test case `ntuple_extended` using perf and g++ 9.3
found that unordered_map::find appeared to be faster than
unordered_map::count. The generated assembly was also shorter for find.

perf output for ROOT::Experimental::RClusterDescriptor::ContainsColumn:

count: 4.67%     4.67%  ntuple_extended  libROOTNTuple.so
find:  3.85%     3.85%  ntuple_extended  libROOTNTuple.so