[tree] Print error in case of branch kind mismatch in CopyAddresses
TTree::CopyAddresses has the built-in pre-condition that the input and
output branches are of the same kind. Clones might be added, however,
for which the pre-condition is violated. This is currently the case,
for example, with certain usages of RDataFrame::Snapshot, which might
create an output branch that is a simple TBranch while the input branch
is e.g. a TBranchElement. This results in wrong data being written out.

With this patch we detect this case and complain.
A proper fix will be proposed soon. The issue is tracked as #8295.