Correct value of TLeaf*::GetDeserializeType and usage thereof.
Rename DeserializeType::kDestructive into DeserializeType::kExternal
as it describe better the situation (the TBufferFile is not large
enough to hold the deserialized data).

Correct the value of GetDeserializeType for TLeafB, TLeafG, TLeafO, TLeafS
in particular several where incorrectly marked as kExternal when
part of an array (fLeafCount != nullptr);

This fixes #6520.