[DF] Fix Book with helper that takes no columns
Recent changes that added support for jitted Book actions broke the
edge case in which the action helper takes no columns:
`df.Book(Helper{}, {})` now means "jit this action" rather than "this
action does not take any columns".

With this patch, we resolve this ambiguity at runtime by looking at
the list of column names. A helper type is used to make sure that
compilation is fine in all cases.