[DF] Include `npartitions` positional argument in head node constructors
The number of partitions in which the distributed RDataFrame should be split is needed in the logic for creating the ranges from the head node, and might also be modified after some checks.
Since it should stay a data member of the head node classes, it should also be defined at __init__ time.
This now happens in the call of `make_dataframe` of any distributed backend, that will also pass the `npartitions` argument to the `get_headnode` factory function.
The `npartitions` parameter is now a required, positional argument in all the signatures that expect it.