[RF] New RooArgList/Set constructors for use in pyROOT
Adding constructors that create a non-owning RooArgLists/Sets from a
vector of RooAbsArg pointers.

This is mainly intended for pyROOT. With cppyy, a Python list or tuple
can be implicitly converted to an `std::vector`, and by enabling implicit
construction of a RooArgList/Set from a `std::vector`, we indirectly
enable implicit conversion from a Python list/tuple to RooArgLists/Sets.