[rdrawable] Introduce ROnFrameDrawable
It should be base class for drawable which can be drawn on frame or
outside frame and which support clipping inside the frame.
Objects which always drawn on frame or which always drawn on the pad,
should not derive from this class. Dedicated for RText, RLince, RBox and
similar objects