[RF] Allow RooArgList/Set construction from vectors containing numbers
Introduces a new struct containing either a `RooAbsArg*` or a `double`
for use in the RooArgList/Set constructors that take a `std::vector`.

The input vector is now a vector of this new struct, meaning we can now
implicitely create RooArgLists/Sets from python collections that contain
either RooAbsArgs, numbers, or both.

With this change Python code that looked like this before:

```Python
ROOT.RooPolynomial("poly", "poly", x,
    ROOT.RooArgList(ROOT.RooConst(1), ROOT.RooConst(2)))
```

can be changed to this:

```Python
ROOT.RooPolynomial("poly", "poly", x, [1, 2])
```