[RF] Add `RooAbsCollection::assign()` as alternative to `operator=()`
The copy-assignment operator of RooAbsCollection, RooArgSet and
RooArgList is confusing because it's not a properly implemented copy
assignment operator, also because it's inconsistent with the copy
constructor. It only copies the values of shared arguments, not the
acual RooAbsCollection. Unfortunately, it is too late to change this
behaviour now.

To avoid bugs and confusion when reading code, an equivalent member
function was introduced: `RooAbsCollection::assign()`. This new method
should be preferred over the copy assignment operator in RooFit code to
make the intention clear that no copy-assignment is intended.