[cxxmodules] Exclude the RInterface from the global module index.
PR root-project/root#6969 allowed ROOT to build pcm files for the R package via
the RInterface.pcm. However, various headers of R define macros which are
defined by math.h and others (ERROR in RooFit).

When the global module index is enabled, ROOT loads RInterface on a lookup for
(Range). Then the exported macros "shadow" the ones which are still in the
index and we do not load the correct corresponding module. This obscure problem
seems to be coming from a current deficiency in the implementation/interaction
between clang and the index.

Pre-loading the RInterface, if enabled, seems to resolve the issue at a
reasonable compromise rather than trying to disable the module altogether or
undefine the fragile macros.