Readjust the representation of CallFunc arguments to match the callee signature.
TClingCallFunc provides an interface between compiled and interpreted code. That
is, we can create at compile time a function to be called by the interpreter and
get back its result into compiled code. To do that we require connecting to a
function declaration available in the interpreter and calling it by setting
input arguments. However, the input arguments are set with the SetArg which can
resolve to a type and argument representation different from the signature of the
interpreter function we will call. This is practically not a problem integral
types because their representation is mostly the same. The problem becomes
visible when we call SetArg with a `double` and the function signature expects
a `float`. That works out of the box when the compiler can see both ends and
insert proper representation casts. Unfortunately, when crossing the
compiler/interpreter boundary we use `void*` and that information is lost.

This patch adds some representation adjustments so that the compiled code
(SetArg) and the interpreted code (CallFunc) can agree on the memory
representation before calling.

Fixes #14405