[geom] Add cache layer for TGeoParallelWorld safeties in TGeoNavigator
This commit provides an (optional) optimization to reduce
the cost of safety evaluations associated with TGeoParallelWorld.

The approach consists in the introduction of additional cache-state
in TGeoNavigator. This state remembers the last location for which
TGeoParallelWorld::Safety was evaluated, together with the corresponding
value returned.

Since TGeoNavigator follows the evolution of tracks step-by-step,
we have some sort of history-locality. This means that the cache-state
has a high-probability of being relevant for multiple track locations in
a row.

The optimization proposed here is orthogonal to other improvements.
It adds on top of algorithmic improvements within TGeoParallelWorld
(bounding volume hierarchies, see separate PR), gaining a few extra percent
in speed.

By default, the new feature is not enabled by default for backward compatibility.
Once fully verified, we can make it the default mode.

That said, in ALICE simulations, the caching did not modify stepping
behaviour and yielded identical results compared to not doing the caching.